# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Magica Clone API",
    "version": "1",
    "description": "Send messages to the agent, read conversations, follow runs, answer approvals, and run Magica tools directly. Every response carries `x-api-version: 1` and `x-trace-id`; every error is `{ error, code, details?, traceId }`. Work that takes time starts at once and returns a run to poll."
  },
  "servers": [
    {
      "url": "https://magica-backend-production.up.railway.app"
    },
    {
      "url": "http://localhost:3000",
      "description": "Running locally (pnpm dev)"
    }
  ],
  "security": [
    {
      "ApiKey": []
    },
    {
      "Bearer": []
    }
  ],
...
  "paths": {
    "/v1/messages": {
      "post": {
        "summary": "Send a message",
        "description": "Sends a message to the agent, in a new chat or (with `chatId`) an existing one, and returns at once with the run to poll.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Makes the request safe to repeat: the same key with the same body (within 24 hours) returns the first answer without doing the work again (`idempotent-replayed: true`); a different body, or a repeat while the first is still running, is a 409 `IDEMPOTENCY_CONFLICT`. A request that failed can be repeated with the same key.",
            "schema": {
              "type": "string",
              "pattern": "^[\\x21-\\x7e]{1,255}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted: the agent is working on it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageAccepted"
                }
              }
            }
          },
          "400": {
            "description": "The request isn't valid (`VALIDATION_FAILED`): `error` says which field and why.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Not enough credits (`INSUFFICIENT_CREDITS`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not yours (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A conflict: a run is already active in the chat (`RUN_ACTIVE`), or the Idempotency-Key was used differently (`IDEMPOTENCY_CONFLICT`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 1 MB (`PAYLOAD_TOO_LARGE`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "A service the request needs is unavailable right now (`SERVICE_UNAVAILABLE`): try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "Create a chat completion",
        "description": "The chat-completions format, answered by the agent through the free model router (`model` must be `openrouter/free`). Not streamed. The conversation is kept as a chat. Waits about a minute for the answer; if it takes longer, answers 202 with the run to poll.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Makes the request safe to repeat: the same key with the same body (within 24 hours) returns the first answer without doing the work again (`idempotent-replayed: true`); a different body, or a repeat while the first is still running, is a 409 `IDEMPOTENCY_CONFLICT`. A request that failed can be repeated with the same key.",
            "schema": {
              "type": "string",
              "pattern": "^[\\x21-\\x7e]{1,255}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The answer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletion"
                }
              }
            }
          },
          "202": {
            "description": "Still working: poll `GET /v1/runs/{run_id}`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCompletionPending"
                }
              }
            }
          },
          "400": {
            "description": "The request isn't valid (`VALIDATION_FAILED`): `error` says which field and why.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Not enough credits (`INSUFFICIENT_CREDITS`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A conflict: a run is already active in the chat (`RUN_ACTIVE`), or the Idempotency-Key was used differently (`IDEMPOTENCY_CONFLICT`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "The body is over 1 MB (`PAYLOAD_TOO_LARGE`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "A service the request needs is unavailable right now (`SERVICE_UNAVAILABLE`): try again.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chats": {
      "get": {
        "summary": "List chats",
        "description": "Your chats, pinned first, then most recent. Pass `cursor` from one page to get the next.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of chats.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatList"
                }
              }
            }
          },
          "400": {
            "description": "The request isn't valid (`VALIDATION_FAILED`): `error` says which field and why.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chats/{chatId}/messages": {
      "get": {
        "summary": "List a chat's messages",
        "description": "Finished messages, oldest to newest within a page; `cursor` fetches older ones. A reply being written is read from its run.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "description": "The chat.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageList"
                }
              }
            }
          },
          "400": {
            "description": "The request isn't valid (`VALIDATION_FAILED`): `error` says which field and why.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not yours (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{runId}": {
      "get": {
        "summary": "Get a run",
        "description": "Where the run stands, its usage, its reply so far, the tool calls it made (with what they made), and what it waits for.",
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "description": "The run.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not yours (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{runId}/cancel": {
      "post": {
        "summary": "Stop a run",
        "description": "Stops the run if it is still going (what it wrote so far is kept), and returns it as it stands.",
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "description": "The run.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not yours (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tools/runs/{runId}": {
      "get": {
        "summary": "Get a tool run",
        "description": "A tool run's status, input, credits, duration, what it made, and its error if it failed.",
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "description": "The tool run.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The tool run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolRunResponse"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not yours (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/credits": {
      "get": {
        "summary": "Get credits",
        "description": "Your balance, and how much of it is held by work in progress.",
        "tags": [
          "Account"
        ],
        "responses": {
          "200": {
            "description": "Your credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credits"
                }
              }
            }
          },
          "401": {
            "description": "No valid API key or session token (`UNAUTHORIZED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "A limit was reached (`RATE_LIMITED`): see `Retry-After` and `details`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "An API key (`mgc_…`) from API / MCP in the app."
      },
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key (`Bearer mgc_…`), or a signed-in session token."
      }
    },
    "schemas": {
...
      "MessageAccepted": {
        "type": "object",
        "properties": {
          "chatId": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "const": "queued"
          },
          "webhook": {
            "type": "object",
            "properties": {
              "signingSecret": {
                "type": "string",
                "format": "starts_with",
                "pattern": "^whsec_.*"
              }
            },
            "required": [
              "signingSecret"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "chatId",
          "messageId",
          "runId",
          "status"
        ],
        "additionalProperties": false
      },
...
    }
  }
}